tests/wpt/css-tests/css-flexbox-1_dev/xhtml1print/flexbox_computedstyle_flex-flow-row-nowrap.xht

<!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Strict//EN" "http://www.w3.org/TR/xhtml1/DTD/xhtml1-strict.dtd">
<html xmlns="http://www.w3.org/1999/xhtml"><head><title>flexbox | computed style | flex-flow: row nowrap</title>
<link href="http://opera.com" rel="author" title="Opera Software" />
<link href="http://www.w3.org/TR/css-flexbox-1/#propdef-flex-flow" rel="help" />
<link href="http://www.w3.org/TR/css-flexbox-1/#valdef-flex-direction-row" rel="help" />
<link href="http://www.w3.org/TR/css-flexbox-1/#valdef-flex-wrap-nowrap" rel="help" />
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<meta content="dom" name="flags" />
<style>
body {
	color: red;
	flex-flow: row nowrap;
}
</style>
</head><body><div id="log"></div>
<script>
test(function() {
    var body = document.body;

    assert_equals(getComputedStyle(body).getPropertyValue("flex-direction"),
	          "row");
    assert_equals(getComputedStyle(body).getPropertyValue("flex-wrap"),
	          "nowrap");
});
</script>
</body></html>
